remove compiler warnings

+++ lua.hpp
@@ -7,6 +7,8 @@
 #include <cstdlib>
 #include <cstdio>
 #include <cstdint>
+#include <cstring>
+#include <memory>
 #include <string>
 #include <set>
 #include <stack>
@@ -110,7 +112,7 @@ class Transpiler {
   // integer literal expression
   struct Integer : public Expression {
     Integer(IntegerLiteral n) : literal(n) { }
-    virtual void transpile(Transpiler& transpiler) { fprintf(transpiler.out, "%llu", literal); }
+    virtual void transpile(Transpiler& transpiler) { fprintf(transpiler.out, "%llu", static_cast<unsigned long long>(literal)); }
     IntegerLiteral literal;
   };
